Fix blocky appearance of water surfaces
The water surface pixel shader implements a kind of dithering effect for
its reflections and refractions that appears like blocks when the
resolution is increased. This can be seen for example in New Londo Ruins
or at the Darkroot Basin. The dithering effect is implemented by
multiplying viewport coordinates with parts of register c164. The
register is set each frame to contain half the expected resolution of
the render buffer in xy, and the reciprocals of the resolution in zw.

This change improves the appearance of the water surface shader by
intercepting updates to register c164 and setting its values to reflect
the actual resolution of the render buffer.