Update water surface blockiness fix
The water surface is rendered into a buffer that has the same resolution
as the first DoF buffer, and the values passed in c164 now reflect this.

The dithering effect is not actually used by the unmodified game, and
values passed through c164 should match the exact resolution of the
render buffer used for the water surface. This is also required if the
water surface is to look good in all choices of DoF buffer resolution.